+++ components/ScheduleOCRModal.tsx
@@ -1,15 +1,14 @@
 'use client';
 
-import { useState, useRef, useEffect } from 'react';
+import { useState, useRef } from 'react';
 import Image from 'next/image';
 import { CameraCapture } from './CameraCapture';
 import { OCRConfirmModal } from './OCRConfirmModal';
-import { extractScheduleFromImage } from '@/lib/scheduleOCR';
 import type { TimeLabel, RoastSchedule } from '@/types';
 import { HiX, HiPhotograph, HiCamera } from 'react-icons/hi';
 import { useToastContext } from './Toast';
-import { useDeveloperMode } from '@/hooks/useDeveloperMode';
 import { useAppData } from '@/hooks/useAppData';
+import { useScheduleImageProcessing } from '@/hooks/useScheduleImageProcessing';
 
 interface ScheduleOCRModalProps {
   selectedDate: string;
@@ -19,100 +18,22 @@ interface ScheduleOCRModalProps {
 
 export function ScheduleOCRModal({ selectedDate, onSuccess, onCancel }: ScheduleOCRModalProps) {
   const { showToast } = useToastContext();
-  const { isEnabled: isDeveloperMode } = useDeveloperMode();
   const { data } = useAppData();
-  const [showCamera, setShowCamera] = useState(false); // 初期状態は選択画面
-  const [isProcessing, setIsProcessing] = useState(false);
-  const [error, setError] = useState<string | null>(null);
+  const [showCamera, setShowCamera] = useState(false);
   const [ocrResult, setOcrResult] = useState<{ timeLabels: TimeLabel[]; roastSchedules: RoastSchedule[] } | null>(null);
   const [showConfirm, setShowConfirm] = useState(false);
-  const [currentFile, setCurrentFile] = useState<File | null>(null);
-  const [imagePreview, setImagePreview] = useState<string | null>(null);
   const fileInputRef = useRef<HTMLInputElement>(null);
 
-  // 画像プレビューを生成
-  useEffect(() => {
-    if (currentFile) {
-      const reader = new FileReader();
-      reader.onload = (e) => {
-        setImagePreview(e.target?.result as string);
-      };
-      reader.readAsDataURL(currentFile);
-    } else {
-      setImagePreview(null);
-    }
-  }, [currentFile]);
-
-  const handleImageFile = async (file: File) => {
-    setCurrentFile(file);
-    setIsProcessing(true);
-    setError(null);
-    setShowCamera(false);
-
-    try {
-      const { timeLabels, roastSchedules } = await extractScheduleFromImage(file, selectedDate);
+  const { isProcessing, error, imagePreview, handleImageFile, resetState } = useScheduleImageProcessing({
+    selectedDate,
+    onSuccess: (timeLabels, roastSchedules) => {
       setOcrResult({ timeLabels, roastSchedules });
       setShowConfirm(true);
-    } catch (err: unknown) {
-      const scheduleError = err as { code?: string; message?: string; details?: unknown };
-      console.error('OCR処理エラー:', err);
-      console.error('エラー詳細:', {
-        code: scheduleError?.code,
-        message: scheduleError?.message,
-        details: scheduleError?.details,
-      });
-      
-      let errorMessage = 'スケジュールの読み取りに失敗しました。画像を確認して再度お試しください。';
-      let errorDetails = '';
-      
-      if (err instanceof Error) {
-        // Firebase Functionsのエラーコードを確認
-        const errorCode = typeof scheduleError?.code === 'string' ? scheduleError.code : '';
-        errorDetails = err.message || '';
-        
-        if (errorCode === 'unauthenticated' || err.message.includes('unauthenticated')) {
-          errorMessage = '認証が必要です。再度ログインしてください。';
-        } else if (errorCode === 'functions/not-found' || err.message.includes('not-found') || err.message.includes('404')) {
-          errorMessage = 'Firebase Functionsが見つかりません。デプロイを確認してください。';
-          errorDetails = 'ocrScheduleFromImage関数がデプロイされているか確認してください。';
-        } else if (errorCode === 'not-found' || err.message.includes('テキストを検出')) {
-          errorMessage = '画像からテキストを検出できませんでした。画像を確認してください。';
-        } else if (errorCode === 'internal' || err.message.includes('internal') || err.message.includes('サーバー')) {
-          errorMessage = 'サーバーエラーが発生しました。';
-          // 詳細なエラーメッセージがある場合は追加
-          if (err.message && !err.message.includes('サーバーエラー')) {
-            errorDetails = err.message;
-          } else {
-            errorDetails = 'Firebase Functionsのログを確認してください。環境変数（GOOGLE_VISION_API_KEY、OPENAI_API_KEY）が設定されているか確認してください。';
-          }
-        } else if (errorCode === 'failed-precondition' || err.message.includes('OPENAI_API_KEY')) {
-          errorMessage = 'サーバー設定エラーが発生しました。';
-          errorDetails = 'OPENAI_API_KEYまたはGOOGLE_VISION_API_KEYが設定されていません。';
-        } else if (errorCode === 'invalid-argument') {
-          errorMessage = '画像データが無効です。';
-          errorDetails = err.message || '';
-        } else if (err.message) {
-          errorMessage = err.message;
-        }
-      }
-      
-      // 開発者モードの場合は詳細なエラー情報を表示
-      const finalErrorMessage = isDeveloperMode && errorDetails 
-        ? `${errorMessage}\n\n詳細: ${errorDetails}`
-        : errorMessage;
-      
-      setError(finalErrorMessage);
-      setShowCamera(false);
-      showToast(errorMessage, 'error');
-      // エラー時はプレビューをクリーンアップ
-      setCurrentFile(null);
-      setImagePreview(null);
-    } finally {
-      setIsProcessing(false);
-    }
-  };
+    },
+  });
 
   const handleCapture = (file: File) => {
+    setShowCamera(false);
     handleImageFile(file);
   };
 
@@ -132,6 +53,13 @@ export function ScheduleOCRModal({ selectedDate, onSuccess, onCancel }: Schedule
     }
   };
 
+  const handleReset = () => {
+    resetState();
+    setShowCamera(false);
+    setOcrResult(null);
+    setShowConfirm(false);
+  };
+
   // 確認画面を表示
   if (showConfirm && ocrResult) {
     const existingTodaySchedule = data?.todaySchedules?.find((s) => s.date === selectedDate);
@@ -146,26 +74,14 @@ export function ScheduleOCRModal({ selectedDate, onSuccess, onCancel }: Schedule
         existingTimeLabels={existingTimeLabels}
         existingRoastSchedules={existingRoastSchedules}
         onSave={(mode, timeLabels, roastSchedules) => {
-          setShowConfirm(false);
-          setOcrResult(null);
-          setCurrentFile(null);
-          setImagePreview(null);
+          handleReset();
           onSuccess(mode, timeLabels, roastSchedules);
         }}
         onCancel={() => {
-          setShowConfirm(false);
-          setOcrResult(null);
-          setCurrentFile(null);
-          setImagePreview(null);
+          handleReset();
           onCancel();
         }}
-        onRetry={() => {
-          setShowConfirm(false);
-          setOcrResult(null);
-          setCurrentFile(null);
-          setImagePreview(null);
-          setShowCamera(false);
-        }}
+        onRetry={handleReset}
       />
     );
   }
@@ -179,8 +95,7 @@ export function ScheduleOCRModal({ selectedDate, onSuccess, onCancel }: Schedule
             <h2 className="text-lg font-semibold text-gray-800">画像を選択</h2>
             <button
               onClick={() => {
-                setCurrentFile(null);
-                setImagePreview(null);
+                handleReset();
                 onCancel();
               }}
               className="p-2 hover:bg-gray-100 rounded-full transition-colors min-h-[44px] min-w-[44px] flex items-center justify-center"
@@ -274,8 +189,7 @@ export function ScheduleOCRModal({ selectedDate, onSuccess, onCancel }: Schedule
             <h2 className="text-lg font-semibold text-gray-800">エラー</h2>
             <button
               onClick={() => {
-                setCurrentFile(null);
-                setImagePreview(null);
+                handleReset();
                 onCancel();
               }}
               className="p-2 hover:bg-gray-100 rounded-full transition-colors min-h-[44px] min-w-[44px] flex items-center justify-center"
